@@ -2,13 +2,27 @@ module evgenykuznetsov.org/go/matebook-applet
 
 require (
 	github.com/BurntSushi/toml v1.2.1
-	github.com/andlabs/ui v0.0.0-20180902183112-867a9e5a498d
+	github.com/andlabs/ui v0.0.0-20200610043537-70a69d6ae31e
 	github.com/cloudfoundry/jibber_jabber v0.0.0-20151120183258-bcc4c8345a21
-	github.com/getlantern/systray v1.2.1
+	github.com/getlantern/systray v1.2.2
 	github.com/nicksnyder/go-i18n/v2 v2.2.1
-	github.com/onsi/ginkgo v1.16.1 // indirect
-	github.com/onsi/gomega v1.11.0 // indirect
-	golang.org/x/text v0.4.0
+	golang.org/x/text v0.9.0
 )
 
-go 1.16
+require (
+	github.com/getlantern/context v0.0.0-20190109183933-c447772a6520 // indirect
+	github.com/getlantern/errors v0.0.0-20190325191628-abdb3e3e36f7 // indirect
+	github.com/getlantern/golog v0.0.0-20190830074920-4ef2e798c2d7 // indirect
+	github.com/getlantern/hex v0.0.0-20190417191902-c6586a6fe0b7 // indirect
+	github.com/getlantern/hidden v0.0.0-20190325191715-f02dbb02be55 // indirect
+	github.com/getlantern/ops v0.0.0-20190325191751-d70cb0d6f85f // indirect
+	github.com/go-stack/stack v1.8.0 // indirect
+	github.com/onsi/ginkgo v1.16.1 // indirect
+	github.com/onsi/gomega v1.11.0 // indirect
+	github.com/oxtoacart/bpool v0.0.0-20190530202638-03653db5a59c // indirect
+	github.com/stretchr/testify v1.5.1 // indirect
+	golang.org/x/sys v0.5.0 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+)
+
+go 1.19
